refactor: change the name of the USPTO input format
Change the name of the patent USPTO input format to show the typical format (XML).
