+++ docling/datamodel/base_models.py
@@ -36,7 +36,7 @@ class InputFormat(str, Enum):
     ASCIIDOC = "asciidoc"
     MD = "md"
     XLSX = "xlsx"
-    PATENT_USPTO = "uspto"
+    XML_USPTO = "uspto"
 
 
 class OutputFormat(str, Enum):
@@ -56,7 +56,7 @@ FormatToExtensions: Dict[InputFormat, List[str]] = {
     InputFormat.IMAGE: ["jpg", "jpeg", "png", "tif", "tiff", "bmp"],
     InputFormat.ASCIIDOC: ["adoc", "asciidoc", "asc"],
     InputFormat.XLSX: ["xlsx"],
-    InputFormat.PATENT_USPTO: ["xml", "txt"],
+    InputFormat.XML_USPTO: ["xml", "txt"],
 }
 
 FormatToMimeType: Dict[InputFormat, List[str]] = {
@@ -83,7 +83,7 @@ FormatToMimeType: Dict[InputFormat, List[str]] = {
     InputFormat.XLSX: [
         "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
     ],
-    InputFormat.PATENT_USPTO: ["application/xml", "text/plain"],
+    InputFormat.XML_USPTO: ["application/xml", "text/plain"],
 }
 
 MimeTypeToFormat = {
